requirements.txt
click==4.0
clipboard==0.0.4
inquirer==2.1.5
PyYAML==3.11
